+++ src/net/JobResult.h
@@ -38,7 +38,7 @@ class JobResult
 {
 public:
     inline JobResult() : poolId(0), diff(0), nonce(0) {}
-    inline JobResult(int poolId, const xmrig::Id &jobId, uint32_t nonce, const uint8_t *result, uint32_t diff) :
+    inline JobResult(int poolId, const xmrig::Id &jobId, uint32_t nonce, const uint8_t *result, uint64_t diff) :
         poolId(poolId),
         diff(diff),
         nonce(nonce),
@@ -73,7 +73,7 @@ public:
 
 
     int poolId;
-    uint32_t diff;
+    uint64_t diff;
     uint64_t nonce;
     uint8_t result[LEN::RESULT];
     xmrig::Id jobId;
